requirements.txt
sqlalchemy
argparse
pyzbar
opencv-python
pillow
argon2-cffi
requests
